requirements.txt
# requirements.txt
Pillow>=9.3.0
